add hack script to run operator locally
the hack script creates a container for the operator
and pushes the container to remote repository and uses
this operator-image to deploy the operator on the local cluster.

p.s. this is intended only for testing purpose.

also, added make targets to run and clean the operator on local cluster