Add git-include-dir flag, update Dockerfile
This commit adds the git-include-dir flag
in container.yaml.in. This will tell Cachito
not to strip the .git folder from archived
WMCO upstream repository.
This commit also updates the Dockerfile for
CPaaS pipeline to sync with the container build
process in CI.This change includes copying of
only specific contents required for the build
inside the container from the source repo.

Change-Id: Ida8aa23a3fb02006f66e7c3fb907435f4fb042be