COPY over .gitignore and move building of WMCO
Related: rhjira#WINC-560
This commit changes Dockerfile and adds the COPY
command to copy the .gitignore file from WMCO
github repository to the container.
This commit also moves the copying of specific files
and building of WMCO binary to the end of the building
binaries block.
This change should fix the addition of a -dirty tag
in the WMCO operator version.

Change-Id: I3185d11f325f4c3dfd0a7b0f0da1aab7ad84715a