[wmco] Modify parseHosts to reuse in CSR controller
This commit modifies the method parseHosts in order
to reuse it in the CSR controller without having
to create a newConfigmapRenconciler object.
It also introduces renaming this method to
parseInstances() for consistent term usage,
moves it to the instances.go file to avoid cyclic
dependency when using it in the csr pkg.